[M68k] Fix requirements and operands for MOVE from/to CCR/SR
This fixes the defined CPU feature requirements for some of the MOVE
from/to CCR/SR instructions. Specifically, MOVE to CCR and MOVE from SR
are available to the whole 68000 family, except MOVE from SR requires
supervisor mode on 68010+.

On 68000, MOVE from SR is used instead of MOVE from CCR to move the
condition bits. A new feature predicate IsM68000 is added for any
lowering that needs to make this choice.

The operands on these instructions are also adjusted to not take a
destination for MOVE to CCR/SR (because the destination is invariable)
and to not take a source for MOVE from CCR/SR (because the source is
invariable).