[M68k] Fix COPY instruction killing live condition flags
During PHI node elimination, a COPY instruction is inserted that is not
known to kill condition flags until after it is lowered by
`copyPhysReg()`. To mitigate this, `copyPhysReg()` now has logic to
determine whether the CCR is live, and if so, will back it up and
restore it after the COPY.

I've only seen this occur due to PHI elimination, but if the root cause
is that LLVM assumes COPY has no side-effects, then perhaps this would
show up elsewhere as well.

[M68k] Fix CCR edge cases in `copyPhysReg()`