[M68k] Implement CLR instruction
The CLR instruction is equivalent to moving immediate 0 to an effective
address, except it's strictly more efficient due to not requiring the
extra bytes and cycles to load the immediate value.

Nodes which set or store 0 will prioritize CLR over MOVE in pattern
matching.

(Note: Usually, when clearing multiple memory addresses, it's faster to
clear a register and store that register multiple times, but this would
have to be achieved separately in an optimization pass.)